requirements.txt
django
djangorestframework
django-filter
djoser 
drf-yasg
python-dotenv